chore: restructure project files and folders
