+++ code/classes/__init__.py
@@ -1,0 +1,2 @@
+from .vehicle import Vehicle, Orientation
+from .board import Board
