+++ data/input/gameboards/Rushhour12x12_7.csv
