+++ data/input/gameboards/Rushhour6x6_1.csv
