+++ data/input/gameboards/Rushhour6x6_2.csv
