+++ data/input/gameboards/Rushhour6x6_3.csv
