+++ data/input/gameboards/Rushhour9x9_6.csv
