+++ data/input/test_boards/Rushhour6x6_test.csv
